pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flights"
version = "0.1.0"
description = "Flights"
requires-python = ">=3.10"
dependencies = [
    "PyYAML", 
    "requests", 
    "pandas", 
    "Jinja2", 
    "joblib"
]

[project.scripts]
flights = "flights.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
